cli: Disable SSL verifications on S3 uploads
I'm using Linode's S3-compatible object storage, with a bucket named
gallery-static.remmy.io so that I can have it proxied through cloudflare
for caching purposes, but mainly so that the images are served from
closer to the user.

The S3 endpoint can't serve a certificate that matches a domain, since
you can't have multi-level wildcard certs.  So the certificate for the
S3 endpoint will never be valid for my uploads.

I wonder if I could somehow build the expected cert into the binary
instead of just disabling verification completely?

Signed-off-by: Johannes Löthberg <johannes@kyriasis.com>